confirmation dialog moved to CountersFragmet
