+++ app/src/main/java/ua/napps/scorekeeper/app/MainActivity.java
@@ -1,7 +1,5 @@
 package ua.napps.scorekeeper.app;
 
-import android.app.AlertDialog;
-import android.content.DialogInterface;
 import android.content.SharedPreferences;
 import android.graphics.Color;
 import android.os.Bundle;
@@ -20,7 +18,6 @@ import com.github.fernandodev.easyratingdialog.library.EasyRatingDialog;
 import ua.com.napps.scorekeeper.R;
 import ua.napps.scorekeeper.counters.CountersFragment;
 import ua.napps.scorekeeper.dice.DicesFragment;
-import ua.napps.scorekeeper.listeners.DialogPositiveClickListener;
 import ua.napps.scorekeeper.settings.LocalSettings;
 import ua.napps.scorekeeper.settings.SettingsFragment;
 import ua.napps.scorekeeper.utils.AndroidFirebaseAnalytics;
@@ -47,7 +44,6 @@ public class MainActivity extends AppCompatActivity implements SharedPreferences
     private boolean isDarkTheme;
     private boolean isKeepScreenOn;
     private BottomNavigationBar bottomNavigationBar;
-    private AlertDialog alertDialog;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -225,28 +221,6 @@ public class MainActivity extends AppCompatActivity implements SharedPreferences
         } else {
             bottomNavigationBar.selectTab(0);
         }
-    }
-
-    public void showDialogWithAction(int message, final DialogPositiveClickListener listener) {
-        if (alertDialog != null && alertDialog.isShowing()) {
-            return;
-        }
-        AlertDialog.Builder builder = new AlertDialog.Builder(this);
-        builder.setMessage(message)
-                .setCancelable(false)
-                .setPositiveButton(R.string.dialog_yes, new DialogInterface.OnClickListener() {
-                    public void onClick(DialogInterface dialog, int id) {
-                        listener.onPositiveButtonClicked(MainActivity.this);
-                    }
-                })
-                .setNegativeButton(R.string.dialog_no, new DialogInterface.OnClickListener() {
-                public void onClick(DialogInterface dialog, int id) {
-                    dialog.dismiss();
-                }
-            });
-
-        alertDialog = builder.create();
-        alertDialog.show();
     }
 
 }
